Update SortingSearching.cpp
Added a Node* deleter that deletes all of the node pointers at the end of the A* algorithm so there wouldn't be a memory leak if you did this process repeatedly. Future ideas besides a file reader would be for the most cost effective path to be printed as well. This might be done with a stack and starting from the end node get each parent node until parent node equals nullptr then run a while loop and print them out with some reasonable syntax for readability